+++ src/sasctl/_services/cas_management.py
@@ -4,7 +4,12 @@
 # Copyright © 2019, SAS Institute Inc., Cary, NC, USA.  All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
+import os
+
 from .service import Service
+
+DEFAULT_SERVER = 'cas-shared-default'
+DEFAULT_CASLIB = 'casuser'
 
 
 class CASManagement(Service):
@@ -53,7 +58,7 @@ class CASManagement(Service):
         RestObj
 
         """
-        server = server or 'cas-shared-default'
+        server = server or DEFAULT_SERVER
         caslibs = cls.list_caslibs(server,
                                    filter='eq($primary,name, "%s")' % name)
 
@@ -84,15 +89,16 @@ class CASManagement(Service):
                             server) or []
 
     @classmethod
-    def get_table(cls, name, caslib, server=None):
+    def get_table(cls, name, caslib=None, server=None):
         """Get a table by name.
 
         Parameters
         ----------
         name : str
             Name of the table.
-        caslib : str or dict
-            Name, ID, or dictionary representation of the caslib.
+        caslib : str or dict, optional
+            Name, ID, or dictionary representation of the caslib.  Defaults to
+            CASUSER.
         server : str, optional
             Server where the `caslib` is registered.
 
@@ -101,9 +107,69 @@ class CASManagement(Service):
         RestObj
 
         """
+        caslib = caslib or DEFAULT_CASLIB
         tables = cls.list_tables(caslib,
                                  server=server,
                                  filter='eq($primary,name, "%s")' % name)
 
         if len(tables) > 0:
             return tables.pop()
+
+    @classmethod
+    def upload_file(cls, file, name, caslib=None, server=None, header=None):
+        """Upload a file to a CAS table.
+
+        Uploads the contents of a CSV, XLS, XLSX, SAS7BDT or SASHDAT file to a
+        newly created CAS table.
+
+        Parameters
+        ----------
+        file : str or file-like object
+            File containing data to upload or path to the file.  Valid file
+            types are .CSV, .XLS, .XSLX, .SAS7BDAT, .SASHDAT.
+        name : str
+            Name of the table to create
+        caslib : str, optional
+            caslib in which the table will be created.  Defaults to CASUSER.
+        server : str, optional
+            CAS server on which the table will be created.  Defaults to
+            cas-shared-default.
+        header : bool, optional
+            Whether the first row of data contains column headers.  Defaults to
+            True.
+
+        Returns
+        -------
+        RestObj
+            Table reference
+
+        """
+        name = str(name)
+        caslib = caslib or DEFAULT_CASLIB
+        server = server or DEFAULT_SERVER
+        header = True if header is None else bool(header)
+
+        # Not a file-like object, assuming it's a file path
+        if not hasattr(file, 'read'):
+            path = os.path.abspath(os.path.expanduser(file))
+            extension = os.path.splitext(path)[-1].lstrip('.').lower()
+
+            # Extension should be supported & needs to be explicitly set in
+            # the "format" parameter to avoid errors.
+            if extension not in ('csv', 'xls', 'xlsx', 'sas7bdat', 'sashdat'):
+                raise ValueError("File '%s' has an unsupported file type." %
+                                 file)
+
+            with open(path, 'rb') as f:
+                file = f.read()
+
+        tbl = cls.post('/servers/%s/caslibs/%s/tables' % (server, caslib),
+                       data={
+                           'tableName': name,
+                           'containsHeaderRow': header,
+                           'format': extension
+                       },
+                       files={
+                           name: file
+                       })
+        return tbl